chore(wren-ui): move data preview files
